requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
asyncpg==0.29.0
databases[postgresql]==0.8.0
sqlalchemy==2.0.23

# Vector Search
pinecone-client==2.2.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# LLM and AI
openai==0.28.1
langchain==0.0.335
tiktoken==0.5.1

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
unstructured[local-inference]==0.10.30
python-magic==0.4.27

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Configuration and Environment
pydantic[email]==2.5.0
python-dotenv==1.0.0

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
